Extensions can be selectively enabled and disabled now. This affects what extension commands look like in a program (e.g. DROP vs. XROM 09,49), whether they are recognized by XEQ, whether they show up in the FCN catalog, and what happens when a program tries to execute them.
There are now 6 extensions: COPAN, Bigstack, ACCEL, LOCAT, HEADING, and HP-41
Time. COPAN is always disabled in Free42 production builds; ACCEL, LOCAT, and
HEADING are enabled by default in the iPhone shell (to do: add an option to the
Preferences dialog to turn them on or off; be sure to call redisplay() when
changing the setting) and always off in all the other shells; Bigstack is
always enabled in 42s and always disabled in all other builds; HP-41 Time is
enabled by default in all shells (and all shells should have an option to
disable this, but none of them do, so far).

Note that HP-41 Time isn't actually implemented yet... Just laying the
groundwork.